src/main/java/com/dphong/musinsa/service/BrandService.java
package com.dphong.musinsa.service;

import com.dphong.musinsa.domain.Brand;
import com.dphong.musinsa.model.request.brand.BrandCreateRequest;
import com.dphong.musinsa.model.response.brand.BrandCreateResponse;
import com.dphong.musinsa.repository.brand.BrandRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class BrandService {

    private final BrandRepository brandRepository;

    public BrandCreateResponse create(BrandCreateRequest request) {
        Brand brand = brandRepository.save(Brand.builder().name(request.name()).build());
        return new BrandCreateResponse(brand.getId(), brand.getName());
    }
}
